+++ hutool-core/src/main/java/com/xiaoleilu/hutool/io/resource/UrlResource.java
@@ -2,7 +2,6 @@ package com.xiaoleilu.hutool.io.resource;
 
 import java.io.BufferedReader;
 import java.io.File;
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 import java.nio.charset.Charset;
@@ -18,7 +17,7 @@ import com.xiaoleilu.hutool.util.URLUtil;
  * @author Looly
  *
  */
-public class UrlResource {
+public class UrlResource implements Resource{
 	protected URL url;
 	
 	//-------------------------------------------------------------------------------------- Constructor start
@@ -33,33 +32,25 @@ public class UrlResource {
 	/**
 	 * 构造
 	 * @param file 文件路径
+	 * @deprecated Please use {@link FileResource}
 	 */
+	@Deprecated
 	public UrlResource(File file) {
 		this.url = URLUtil.getURL(file);
 	}
 	//-------------------------------------------------------------------------------------- Constructor end
 	
-	/**
-	 * 获得解析后的{@link URL}
-	 * @return 解析后的{@link URL}
-	 */
-	public final URL getUrl(){
+	@Override
+	public URL getUrl(){
 		return this.url;
 	}
 	
-	/**
-	 * 获得 {@link InputStream}
-	 * @return {@link InputStream}
-	 */
+	@Override
 	public InputStream getStream(){
 		if(null == this.url){
 			throw new IORuntimeException("Resource [{}] not exist!", this.url);
 		}
-		try {
-			return this.url.openStream();
-		} catch (IOException e) {
-			throw new IORuntimeException(e);
-		}
+		return URLUtil.getStream(url);
 	}
 	
 	/**
@@ -69,19 +60,11 @@ public class UrlResource {
 	 * @since 3.0.1
 	 */
 	public BufferedReader getReader(Charset charset){
-		return IoUtil.getReader(getStream(), charset);
+		return URLUtil.getReader(this.url, charset);
 	}
 	
 	//------------------------------------------------------------------------------- read
-	/**
-	 * 读取资源内容，读取完毕后会关闭流<br>
-	 * 关闭流并不影响下一次读取
-	 * 
-	 * @param charset 编码
-	 * @return 读取资源内容
-	 * @throws IORuntimeException 包装{@link IOException}
-	 * @since 3.0.8
-	 */
+	@Override
 	public String readStr(Charset charset) throws IORuntimeException{
 		BufferedReader reader = null;
 		try {
@@ -92,26 +75,12 @@ public class UrlResource {
 		}
 	}
 	
-	/**
-	 * 读取资源内容，读取完毕后会关闭流<br>
-	 * 关闭流并不影响下一次读取
-	 * 
-	 * @return 读取资源内容
-	 * @throws IORuntimeException 包装{@link IOException}
-	 * @since 3.0.8
-	 */
+	@Override
 	public String readUtf8Str() throws IORuntimeException{
 		return readStr(CharsetUtil.CHARSET_UTF_8);
 	}
 	
-	/**
-	 * 读取资源内容，读取完毕后会关闭流<br>
-	 * 关闭流并不影响下一次读取
-	 * 
-	 * @return 读取资源内容
-	 * @throws IORuntimeException 包装{@link IOException}
-	 * @since 3.0.8
-	 */
+	@Override
 	public byte[] readBytes() throws IORuntimeException{
 		InputStream in = null;
 		try {